UC-De-048: Data_length alterada de 4 bytes para 2 bytes

+++ SimuladorMaquina/base/core/comunicacao/informacao_comunicacao.h
@@ -22,7 +22,7 @@ typedef struct Payload {
     unsigned char version;
     unsigned char code;
     unsigned short id;
-    unsigned int data_length;
+    unsigned short data_length;
     char *data;
 } Payload;
